Fixes for windows warnings savannah 15808
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@38139 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: f4e19c6fbe9d04fcd2d96fafa03de34d634bb6e3

+++ Rec/Muon/MuonTools/src/MuonFastPosTool.cpp
@@ -77,8 +77,8 @@ StatusCode MuonFastPosTool::initialize() {
             StatusCode sc =  m_muonDetector->Tile2XYZ(tile,xp,dx,yp,dy,zp,dz);
             if (!sc) return sc;
             (m_pos[station])[index++]=Gaudi::XYZPoint(xp,yp,zp);
-            m_padSizeX[station*4+region]=dx;
-            m_padSizeY[station*4+region]=dy;
+            m_padSizeX[station*4+region]=(float) dx;
+            m_padSizeY[station*4+region]=(float) dy;
           }          
         }
         for(unsigned y=m_padGridY[station];y<2*m_padGridY[station];y++){
